Code maintenance

+++ Ion/Source/Core/Platform/Windows/WindowsInput.h
@@ -29,7 +29,7 @@ namespace Ion
 			return 0;
 		}
 
-		static constexpr uint s_InputKeyCodeLookup[256] = {
+		static constexpr byte s_InputKeyCodeLookup[256] = {
 			// 0x00 ---------------------------
 			0x00, // 0x00 : Invalid keycode
 			Mouse::Left,
